1.3.63.1000.alpha.1.aix

+++ src/main/java/tech/ravon/model/iviep/Course.java
@@ -28,7 +28,7 @@ import lombok.Data;
 public class Course implements Serializable {
 
     @TableId(type = IdType.AUTO)
-    private Integer courseId;
+    private Long courseId;
 
     private String courseName;
